chore: fix code style (#588)
