+++ build.gradle
@@ -56,7 +56,7 @@ allprojects {
     test {
         useJUnitPlatform()
     }
-
+    apply from: rootProject.file('gradle/checkstyle.gradle')
 }
 
 configure(subprojects.findAll { it.name != 'tests' }) {
@@ -69,17 +69,17 @@ configure(subprojects.findAll { it.name != 'tests' }) {
     apply plugin: 'com.jfrog.bintray'
 
     task javadocJar(type: Jar) {
-        classifier = 'javadoc'
+        archiveClassifier.set('javadoc')
         from javadoc
     }
 
     task sourcesJar(type: Jar) {
-        classifier = 'sources'
+        archiveClassifier.set('sources')
         from sourceSets.main.allSource
     }
 
     task testJar(type: Jar) {
-        classifier = 'tests'
+        archiveClassifier.set('test-sources')
         from sourceSets.test.output
     }
