+++ ckb-mercury-sdk/build.gradle
@@ -10,9 +10,6 @@ dependencies {
     testCompile("org.junit.platform:junit-platform-runner:1.9.0")
 }
 
-apply from: rootProject.file('gradle/checkstyle.gradle')
-
-
 test {
     useJUnitPlatform()
 }
