+++ ckb-mercury-sdk/src/test/java/mercury/BuildSimpleTransferTransactionTest.java
@@ -17,8 +17,8 @@ public class BuildSimpleTransferTransactionTest {
     SimpleTransferPayloadBuilder builder = new SimpleTransferPayloadBuilder();
     builder.assetInfo(AssetInfo.newCkbAsset());
     builder.addFrom("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqfqyerlanzmnkxtmd9ww9n7gr66k8jt4tclm9jnk");
-    builder.addTo("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqg958atl2zdh8jn3ch8lc72nt0cf864ecqdxm9zf"
-        , BigInteger.valueOf(10000000000L));
+    builder.addTo("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqg958atl2zdh8jn3ch8lc72nt0cf864ecqdxm9zf",
+                  BigInteger.valueOf(10000000000L));
     builder.feeRate(500L);
 
     TransactionWithScriptGroups s =
