+++ ckb-mercury-sdk/src/test/java/mercury/BuildTransferTransactionTest.java
@@ -19,8 +19,8 @@ public class BuildTransferTransactionTest {
     TransferPayloadBuilder builder = new TransferPayloadBuilder();
     builder.assetInfo(AssetInfo.newCkbAsset());
     builder.addFrom(ItemFactory.newAddressItem("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqfqyerlanzmnkxtmd9ww9n7gr66k8jt4tclm9jnk"));
-    builder.addTo("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqg958atl2zdh8jn3ch8lc72nt0cf864ecqdxm9zf"
-        , BigInteger.valueOf(100));
+    builder.addTo("ckt1qzda0cr08m85hc8jlnfp3zer7xulejywt49kt2rr0vthywaa50xwsqg958atl2zdh8jn3ch8lc72nt0cf864ecqdxm9zf",
+                  BigInteger.valueOf(100));
     builder.payFee(TransferPayload.PayFee.FROM);
     builder.feeRate(1100L);
