+++ ckb/build.gradle
@@ -6,5 +6,4 @@ dependencies {
     compile project(":serialization")
 }
 
-apply from: rootProject.file('gradle/checkstyle.gradle')
 apply plugin: 'com.github.johnrengelman.shadow'
