+++ ckb/src/main/java/org/nervos/ckb/transaction/DaoTransactionBuilder.java
@@ -3,7 +3,7 @@ package org.nervos.ckb.transaction;
 import org.nervos.ckb.Network;
 import org.nervos.ckb.service.Api;
 import org.nervos.ckb.sign.TransactionWithScriptGroups;
-import org.nervos.ckb.transaction.scriptHandler.ScriptHandler;
+import org.nervos.ckb.transaction.handler.ScriptHandler;
 import org.nervos.ckb.type.*;
 import org.nervos.ckb.utils.MoleculeConverter;
 import org.nervos.ckb.utils.Numeric;
@@ -14,7 +14,7 @@ import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.Iterator;
 
-import static org.nervos.ckb.transaction.scriptHandler.DaoScriptHandler.*;
+import static org.nervos.ckb.transaction.handler.DaoScriptHandler.*;
 
 public class DaoTransactionBuilder extends AbstractTransactionBuilder {
   CkbTransactionBuilder builder;
@@ -48,6 +48,8 @@ public class DaoTransactionBuilder extends AbstractTransactionBuilder {
       case CLAIM:
         builder.reward += getDaoReward(daoOutPoint);
         break;
+      default:
+        throw new IllegalArgumentException("Unsupported transaction type");
     }
     builder.transactionInputs.add(input);
   }
@@ -150,8 +152,8 @@ public class DaoTransactionBuilder extends AbstractTransactionBuilder {
     }
     CellOutput output = new CellOutput(
         depositCellCapacity,
-        Address.decode(address).getScript()
-        , DAO_SCRIPT);
+        Address.decode(address).getScript(),
+        DAO_SCRIPT);
     byte[] data = MoleculeConverter.packUint64(depositBlockNumber).toByteArray();
     builder.addOutput(output, data);
     return this;
