+++ ckb/src/main/java/org/nervos/ckb/transaction/handler/SudtScriptHandler.java
@@ -1,4 +1,4 @@
-package org.nervos.ckb.transaction.scriptHandler;
+package org.nervos.ckb.transaction.handler;
 
 import org.nervos.ckb.Network;
 import org.nervos.ckb.sign.ScriptGroup;
