+++ config/checkstyle/checkstyle.xml
@@ -15,7 +15,8 @@
     Authors: Max Vetrenko, Ruslan Diachenko, Roman Ivanov.
  -->
 
-<module name = "Checker">
+<module name="Checker">
+    <module name="SuppressWarningsFilter"/>
 
     <property name="charset" value="UTF-8"/>
 
@@ -212,5 +213,6 @@
             <property name="exceptionVariableName" value="expected"/>
         </module>
         <module name="CommentsIndentation"/>
+        <module name="SuppressWarningsHolder"/>
     </module>
 </module>
