+++ core/src/test/java/type/ScriptTest.java
@@ -14,6 +14,7 @@ import java.math.BigInteger;
 public class ScriptTest {
 
   @Test
+  @SuppressWarnings("checkstyle:linelength")
   public void testScriptHashWithCodeHash() throws IOException {
     byte[] code = Numeric.hexStringToByteArray("0x1400000000000e00100000000c000800000004000e0000000c00000014000000740100000000000000000600080004000600000004000000580100007f454c460201010000000000000000000200f3000100000078000100000000004000000000000000980000000000000005000000400038000100400003000200010000000500000000000000000000000000010000000000000001000000000082000000000000008200000000000000001000000000000001459308d00573000000002e7368737472746162002e74657874000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000b000000010000000600000000000000780001000000000078000000000000000a000000000000000000000000000000020000000000000000000000000000000100000003000000000000000000000000000000000000008200000000000000110000000000000000000000000000000100000000000000000000000000000000000000");
     byte[] codeHash = Blake2b.digest(code);
