+++ example/build.gradle
@@ -7,5 +7,3 @@ dependencies {
     compile "com.google.code.gson:gson:$gsonVersion"
     compile "org.slf4j:slf4j-api:$slf4jVersion"
 }
-
-apply from: rootProject.file('gradle/checkstyle.gradle')
