+++ utils/build.gradle
@@ -4,5 +4,3 @@ dependencies {
     compile "org.bouncycastle:bcprov-jdk15on:$bouncycastleVersion"
     compile "com.google.guava:guava:$guavaVersion"
 }
-
-apply from: rootProject.file('gradle/checkstyle.gradle')
